requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
google-cloud-storage
joblib
pyarrow
fastparquet